% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_events.R
\name{combined_events}
\alias{combined_events}
\title{Combined events results}
\usage{
combined_events(marks, scores, event_names, event, seconds, ...)
}
\arguments{
\item{...}{}
}
\value{
An object of class "\code{combined_events}"
}
\description{
\code{combined_events()} is a generic function used to present results
of calls to \code{\link[=scores]{scores()}} and \code{\link[=marks]{marks()}}.
}
