% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedevents-package.R
\docType{package}
\name{combinedevents-package}
\alias{combinedevents-package}
\title{combinedevents: Calculate Scores and Marks for Track and Field
Combined Events}
\description{
The package includes functions to calculate scores and marks for track and
field combined events competitions. The functions are based on the scoring
tables for combined events set forth by the International Association of
Athletics Federation (2001).
}
\references{
International Association of Athletics Federation (2001).
\emph{IAAF Scoring Tables for Combined Events}.
}
\seealso{
Useful links:
\itemize{
\item \url{https://katie-frank.github.io/combinedevents/}
\item \url{https://github.com/katie-frank/combinedevents}
\item Report bugs at \url{https://github.com/katie-frank/combinedevents/issues}
}
}
\author{
\strong{Maintainer}: Katie Frank \email{katiexfrank@gmail.com}
}
