% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinedevents.R
\docType{package}
\name{combinedevents}
\alias{combinedevents}
\title{combinedevents: Calculate Scores and Marks for Track and Field
Combined Events}
\description{
Includes functions to calculate scores and marks for track and
field combined events competitions. The functions are based on the IAAF
Scoring Tables for Combined Events.
}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/katie-frank/combinedevents}
\item Report bugs at \url{https://github.com/katie-frank/combinedevents/issues}
}
}
\author{
\strong{Maintainer}: Katie Frank \email{katiexfrank@gmail.com}
}
